ace/pyproject.toml
[project]
name = "ace"
version = "0.1.0"
description = "Autopilot Content Engine tooling for RapidAI"
authors = [
  { name = "RapidAI", email = "engineering@rapiai.ai" }
]
requires-python = ">=3.11"
dependencies = [
  "apscheduler>=3.10",
  "langchain==0.2.6",
  "langgraph==0.0.40",
  "sqlalchemy>=2.0",
  "pydantic>=2.7",
  "qdrant-client>=1.8",
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "ruamel.yaml>=0.18",
  "structlog>=24.1",
  "loguru>=0.7",
  "feedparser>=6.0",
  "beautifulsoup4>=4.12",
  "jinja2>=3.1",
  "textstat>=0.7",
  "uvloop; platform_system != 'Windows'",
  "tenacity>=8.2"
]

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"
